move fast path to separate file
